requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
tensorflow-gpu>=2.12.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web demo
streamlit>=1.25.0
gradio>=3.35.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0

# Model serving
onnx>=1.14.0
onnxruntime>=1.15.0

# Logging and monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0 